C#: Re-factor the model generator tests.
